+++ src/pararnn/cells/__init__.py
@@ -1,3 +1,4 @@
+from pararnn.cells.para_cfc import ParaCfC
 from pararnn.cells.para_gru import ParaGRU
 from pararnn.cells.para_lstm import ParaLSTM
 from pararnn.cells.para_m2rnn import ParaM2RNN
@@ -6,6 +7,7 @@ from pararnn.cells.para_slstm import ParaSLSTM
 from pararnn.cells.protocol import RNNCell, check_cell
 
 __all__ = [
+    "ParaCfC",
     "ParaGRU",
     "ParaLSTM",
     "ParaM2RNN",